python-requirements.txt
protobuf
sqlalchemy
Flask
Flask-RESTful
PyYAML
